* 'feature/Q&A' of https://github.com/zhiyicx/thinksns-plus-android: (47 commits)
  add  tools namespace
  (#559) 添加切图文件
  (#555)   code   update  follow state  in dynamic detail
  merge  data
  (#555)    code   update  follow  request  target user id
  (#555)    code   update  follow  logice in dynamic detail
  (#558) code 我收藏的圈子中的动态数据转换
  (#555)   code   update  gruop list  show logic
  (#558) code 我收藏的圈子中的动态完成
  (#555)  code   update group dynamic detail  digglist show logic
  (#555)  code  update diguserinfo  in dynamic detial
  (#535) code 圈子他人动态不能收藏bug修复
  (#555)   code  update user info  for circle
  (#558) code 圈子-获取收藏的动态完成
  (#557)    code      DynamicDigListBean   add  Serializable
  (#557)   code  update  DynamicCommentBeanV2
  (#451) code 动态详情用户头像v1bug修复
  (#558) code 圈子-获取收藏的动态列表
  (#557)  code  check empty comment list  data
  (#557)  code  update  comment bean
  ...

# Conflicts:
#	app/src/main/res/layout/fragment_qa.xml

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/SystemRepository.java
@@ -339,7 +339,7 @@ public class SystemRepository implements ISystemRepository {
                     .subscribe(new BaseSubscribe<Conversation>() {
                         @Override
                         protected void onSuccess(Conversation data) {
-                            data.setIm_uid(AppApplication.getmCurrentLoginAuth().getUser_id());
+                            data.setIm_uid((int) AppApplication.getmCurrentLoginAuth().getUser_id());
                             data.setUsids(uids);
                             data.setPair(pair);
                             mChatRepository.insertOrUpdateConversation(data);
